Add annotation support for record classes and refactoring of annotation functionality.
