+++ src/main/java/net/datafaker/annotaions/FieldFake.java
@@ -1,0 +1,19 @@
+package net.datafaker.annotaions;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.FIELD, ElementType.RECORD_COMPONENT})
+public @interface FieldFake {
+
+    String expression() default "";
+
+    String method() default "";
+
+    String languageTag() default "en";
+
+    long[] seed() default {};
+}
