+++ src/main/java/net/datafaker/annotaions/ObjectProvider.java
@@ -1,0 +1,6 @@
+package net.datafaker.annotaions;
+
+interface ObjectProvider {
+
+    <K> K createObject(Class<K> clazz);
+}
