+++ src/main/java/net/datafaker/annotaions/Provider.java
@@ -1,19 +1,8 @@
 package net.datafaker.annotaions;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.reflect.AnnotatedElement;
 
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface Provider {
+interface Provider<R extends AnnotatedElement> {
 
-    String expression() default "";
-
-    String method() default "";
-
-    String languageTag() default "en";
-
-    long[] seed() default {};
+    Object getValue(R field);
 }
